Changesdoneee

+++ src/main/java/com/Spring/LibraryService/Service/BookService.java
@@ -1,10 +1,12 @@
 package com.Spring.LibraryService.Service;
+import com.Spring.LibraryService.entity.BookIsbn;
 import com.Spring.LibraryService.entity.LibraryBook;
 import com.Spring.LibraryService.Model.Book;
 import com.Spring.LibraryService.Model.BookIsbnDTO;
 import com.Spring.LibraryService.Model.CheckoutDTO;
 import com.Spring.LibraryService.Repository.LibraryBookRepository;
 import jakarta.persistence.criteria.CriteriaBuilder;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,13 +20,47 @@ public class BookService {
     //private Map<Long, Book> bookMap = new HashMap<>();
     @Autowired
     private LibraryBookRepository bookRepository;
-    public LibraryBook createBook(LibraryBook book){
+    public BookService(LibraryBookRepository bookRepository) {
+        this.bookRepository = bookRepository;
+    }
+    public Book createBook(Book book){
 
-        // call the database
-        //Integer bookId = new Random().nextInt();
-        //book.setBookId(bookId);
-        //bookMap.put(bookId, book);
-        return  bookRepository.save(book);
+        LibraryBook libraryBook = new LibraryBook();
+
+        // Manually map properties from Book to LibraryBook
+        BeanUtils.copyProperties(book, libraryBook);
+
+        // Map BookIsbn objects from Book to LibraryBook
+        List<BookIsbn> bookIsbns = book.getBookIsbns().stream().map(bi -> {
+            BookIsbn bookIsbn = new BookIsbn();
+            BeanUtils.copyProperties(bi, bookIsbn);
+            bookIsbn.setLibraryBook(libraryBook); // Set LibraryBook reference for each BookIsbn
+            return bookIsbn;
+        }).collect(Collectors.toList());
+
+        // Set the BookIsbn list to the LibraryBook
+        libraryBook.setBookIsbns(bookIsbns);
+
+        // Save the LibraryBook to the database
+        LibraryBook savedLibraryBook = bookRepository.save(libraryBook);
+
+        // Create a new Book object to return
+        Book savedBook = new Book();
+
+        // Manually map properties from saved LibraryBook to Book
+        BeanUtils.copyProperties(savedLibraryBook, savedBook);
+
+        // Manually map BookIsbn objects from saved LibraryBook to Book
+        List<BookIsbnDTO> savedBookIsbns = savedLibraryBook.getBookIsbns().stream().map(savedBi -> {
+            BookIsbnDTO savedBiDto = new BookIsbnDTO();
+            BeanUtils.copyProperties(savedBi, savedBiDto);
+            return savedBiDto;
+        }).collect(Collectors.toList());
+
+        // Set the BookIsbnDTO list to the Book
+        savedBook.setBookIsbns(savedBookIsbns);
+
+        return savedBook;
     }
 
     public Book getBook(Integer bookId) {
@@ -48,22 +84,62 @@ public class BookService {
                     return bdo;
                 }).collect(Collectors.toList());
 
-        /*BookIsbnDTO bookIsbnDTO = new BookIsbnDTO();
-        bookIsbnDTO.setBookIsbn(libraryBook.getBookIsbn().getIsbn());
-        bookIsbnDTO.setBookId(libraryBook.getBookIsbn().getBookId());
-*/
+
         book.setBookIsbns(bookIsbnDTOS);
 
         return book;
     }
 
-    public LibraryBook updateBook(LibraryBook LibraryBook) {
+    public Book updateBook(Book book) {
         //Integer bookId = LibraryBook.getBookId();
         //bookMap.put(bookId, book);
-        return bookRepository.save(LibraryBook);
+        //return bookRepository.save(LibraryBook);
+
+        // Create a new LibraryBook object
+        LibraryBook libraryBook = new LibraryBook();
+
+        // Manually map properties from Book to LibraryBook
+        BeanUtils.copyProperties(book, libraryBook);
+
+        // Map BookIsbn objects from Book to LibraryBook
+        List<BookIsbn> bookIsbns = book.getBookIsbns().stream().map(bi -> {
+            BookIsbn bookIsbn = new BookIsbn();
+            BeanUtils.copyProperties(bi, bookIsbn);
+            bookIsbn.setLibraryBook(libraryBook); // Set LibraryBook reference for each BookIsbn
+            return bookIsbn;
+        }).collect(Collectors.toList());
+
+        // Set the BookIsbn list to the LibraryBook
+        libraryBook.setBookIsbns(bookIsbns);
+
+        // Save the LibraryBook to the database
+        LibraryBook savedLibraryBook = bookRepository.save(libraryBook);
+
+        // Create a new Book object to return
+        Book savedBook = new Book();
+
+        // Manually map properties from saved LibraryBook to Book
+        BeanUtils.copyProperties(savedLibraryBook, savedBook);
+
+        // Manually map BookIsbn objects from saved LibraryBook to Book
+        List<BookIsbnDTO> savedBookIsbns = savedLibraryBook.getBookIsbns().stream().map(savedBi -> {
+            BookIsbnDTO savedBiDto = new BookIsbnDTO();
+            BeanUtils.copyProperties(savedBi, savedBiDto);
+            return savedBiDto;
+        }).collect(Collectors.toList());
+
+        // Set the BookIsbnDTO list to the Book
+        savedBook.setBookIsbns(savedBookIsbns);
+
+        return savedBook;
     }
 
     public void deleteBook(Integer bookId) {
-        bookRepository.deleteById(bookId);
+        Optional<LibraryBook> bookOptional = bookRepository.findById(bookId);
+        if (bookOptional.isPresent()) {
+            bookRepository.deleteById(bookId);
+        } else {
+            throw new IllegalArgumentException("Book with ID " + bookId + " not found");
+        }
     }
 }
